Complete Deletion of all webClient Components due to a switch to the StudIP-API. - Mass of Refactoring - Implementation of Login-Request
