+++ src/studip/api/types/News.java
@@ -1,4 +1,4 @@
-package objects;
+package studip.api.types;
 
 /**
  * The type News.
