+++ src/utils/WebClient.java
@@ -1,0 +1,15 @@
+package utils;
+
+import java.io.File;
+import java.net.URL;
+
+public class WebClient {
+
+
+    public WebClient() {
+
+    }
+
+    public void DownloadFile(URL link, File output) {
+    }
+}
